# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CSE341 Event Planner API",
    "description": "REST API for managing events and venues with MongoDB, validation, error handling, and GitHub OAuth.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "GitHubOAuthSession": {
      "type": "apiKey",
      "in": "cookie",
      "name": "connect.sid",
      "description": "Session cookie created after logging in at /auth/github."
    }
  },
  "paths": {
...
    "/auth/github": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Start GitHub OAuth login",
        "description": "Redirects the user to GitHub to begin OAuth login.",
        "responses": {
          "302": {
            "description": "Redirect to GitHub OAuth"
          },
          "500": {
            "description": "GitHub OAuth is not configured",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "GitHub OAuth callback",
        "description": "GitHub redirects here after successful login. The app stores the OAuth user and creates a session.",
        "responses": {
          "302": {
            "description": "Redirect to Swagger docs after login"
          },
          "401": {
            "description": "Login was not completed",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "OAuth login failed",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out",
        "description": "Ends the current login session.",
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "500": {
            "description": "Logout failed",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "error": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Event not found"
                }
              }
            }
          }
        }
      }
    },
...
  }
}